Redesigned fnProcessing().
The function fnProcessing() was renamed to ino_to_cpp(). Its interface
has changed so that it no longer receiving a file to preprocess.
Instead, it takes the content of the file and returns the output. A new
wrapper function might be added in the future to handle I/O.

In the end, this should not be part of the compiler and must be moved to
another subsystem.