compiler.php: use GCC for syntax checking.
A source file will always be compiled with GCC. If that fails, it will
be then compiled again with Clang to provide better error reporting.
Otherwise, Clang is not used.

This was necessary to address some issues where GCC would compile
successfully but Clang would fail due to the AVR architecture (e.g.
8-bit integers).