Moved configuration options to a new file.
Introduced a new file where all configuration options are stored. Having
deployment-specific options inside the source code did not make much
sense. This is particularly important for people who want to perform
their own installations of codebender.

Currently, 'config.php-dist' should be copied to 'config.php' and edited
by the user. In the future, the configuration may become global for all
codebender and be placed outside of the compiler.