Add support for assembler code.
Some libraries contain code written in assembly. They are now supported
just as C and C++ files.

Please note that .inc files are treated like .h files in function
create_objects(). Though they are included in other files, they are not
headers. The documentation and/or variable names might have to be
updated in the future to avoid confusion.