Fixed input validation bug
Adding the setLoggingParams function in the DefaultController enabled
the option of making log files during compilation, however it also
caused a bug in the validation of the request.  Moving the function's
definition to the CompilerHandler class fixed the bug. We also had to
add a flag (initCall) to the CompilerHandler's main function so that,
when called for the first time, if needed, the compiler makes the log
file. During the next recursive calls of the CompilerHandler class, no
new log files are made, instead, we commands executed are written to the
initial log file.