Handle project mongo id, when provided
This is added for library testing. The x-mongo-id header is a custom
HTTP header sent to the compiler during library testing compilations.
If the header exists, the compiler will send the mongo project id back
to the requestor. This is a bit ugly, but it's the only way to identify
projects during batch compilation attempts.