CompilerV2Handler: add v2 compiler
The "v2" compiler uses the new arduino-builder, introduced by
Arduino in version 1.5.  It wraps the build system in a go-based
backend.  This backend is used by the desktop application, so
it will have the same compatibility and performance as the
Arduino IDE provides.

Boards should be specified with a "fqbn" property, such as:

    * arduino:sam:arduino_due_x
    * arduino:avr:uno

Currently, the following are unsupported:

    * "autocomplete" output
    * "library" compilation

Additionally, "binary" or "hex" output may be unavailable, depending
on which platform you're using.  This is because AVR provides .hex
files by default, and ARM provides .bin files.

Signed-off-by: Sean Cross <sean@xobs.io>