CompilerV2Handler: remove "format" build option
The "format" build option allowed for retreiving the output file
in a variety of formats, from binary to hex to elf.  There were
even a few legacy formats such as "object" and "syntax".

These are no longer supported, and so we remove this option to clean
up dead code paths.

Signed-off-by: Sean Cross <sean@xobs.io>