Convert output to base64 in all compilation cases
The type of the actual compilation output (binary, ihex, etc) depends on
the platform configuration. For instance, for ARM devices the output is
an ihex file, whereas for SAM devices it's a binary file.
Since we can no more request for a specific type of output (like in the
1.0.5 compiler) we decided to always convert the output to base64
encoded text before sending it back to the requestor.